+++ packages/NLP/CHANGELOG.md
@@ -1,5 +1,17 @@
 # NLP changelog
 
+## 1.0.7 (WIP)
+
+Stemming-based search (SBS) is implemented.
+
+### Features
+
+* SBS context panel for text columns
+* SBS performance research tools
+
+*Dependency: datagarok-api >= 1.0.0*
+
+
 ## 1.0.6 (2023-07-24) 
 
 Initial release of NLP package